Commit the source code for the cross compilable reflective dll injection module. Some minor modifications to the stdapi extension were also required. All the projects (.vcproj) now have an x64 debug/release target as well as an x86 counterpart.
git-svn-id: file:///home/svn/framework3/trunk@7000 4d416f70-5f16-0410-b530-b9f4589650da

+++ external/source/meterpreter/source/extensions/stdapi/server/sys/process/util.c
@@ -78,7 +78,7 @@ DWORD copy_memory_to_process(HANDLE process, BOOLEAN allocate,
 		LPVOID *buffer, DWORD length, DWORD prot)
 {
 	LPVOID remoteBuffer = *buffer;
-	DWORD  written;
+	SIZE_T  written;
 	DWORD  result = ERROR_SUCCESS;
 
 	do
